requirements.txt
# Core dependencies
anthropic>=0.18.0
openai>=1.12.0
pyyaml>=6.0
numpy>=1.24.0

# Speech services
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Optional TTS providers
# elevenlabs>=0.2.0
# TTS>=0.22.0  # Coqui TTS

# Web services
gradio>=4.19.0
aiohttp>=3.9.0
websockets>=12.0

# Optional performance
# deepgram-sdk>=3.0.0  # For Deepgram STT

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
